+++ doc/CHANGES.md
@@ -7,7 +7,7 @@ Please see [rsbackup in git](https://github.com/ewxrjk/rsbackup) for detailed ch
 * The names of hosts, volumes, devices and stores are now ordered in a way that respects integer values embedded in them. Fixes [issue #76](https://github.com/ewxrjk/rsbackup/issues/76).
 * A new `rsync-remote` directive allows control over the remote rsync command.
 * The pruning log in the backup report is now displayed in a less verbose form. Fixes [issue #69](https://github.com/ewxrjk/rsbackup/issues/69).
-* Warnings about unknown volumes are more informative. Fixes [issue #79](https://github.com/ewxrjk/rsbackup/issues/79).
+* Warnings about unknown hosts and volumes are more informative. Fixes [issue #79](https://github.com/ewxrjk/rsbackup/issues/79).
 
 ## Changes In rsbackup 7.0